feat/overlay-e2e-config: fixed MR comments

+++ apps/chat-e2e/src/hooks/global-setup.ts
@@ -1,5 +1,6 @@
 import { ResultFolder } from '@/src/testData';
 import { FileUtil } from '@/src/utils';
+import type { FullConfig } from '@playwright/test';
 import path from 'path';
 
 export const isApiStorageType =
@@ -18,9 +19,11 @@ export const ExecutionResults = {
   testResultsPath: path.resolve(__dirname, `../../${ResultFolder.testResults}`),
 };
 
-async function globalSetup() {
-  for (const path of Object.values(ExecutionResults)) {
-    FileUtil.deleteFolder(path);
+async function globalSetup(config: FullConfig) {
+  if (config.projects[0].name === 'auth') {
+    for (const path of Object.values(ExecutionResults)) {
+      FileUtil.deleteFolder(path);
+    }
   }
 }
